refactor: extract inline tests to separate files and add coverage tests
Move inline #[cfg(test)] blocks from wardnet-types source files (auth.rs,
device.rs, event.rs, routing.rs) into dedicated test files under src/tests/,
matching the project convention. Add device service tests for admin-locked,
anonymous, and not-found edge cases. Add middleware tests for device MAC
resolution path.
