+++ source/daemon/crates/wardnet-types/src/tests/mod.rs
@@ -1,4 +1,8 @@
 mod api;
+mod auth;
+mod device;
+mod event;
+mod routing;
 mod tunnel;
 mod vpn_provider;
 mod wireguard_config;
